web app/requirements.txt
absl-py==0.12.0
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-WTF==0.14.3
gast==0.3.3
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.37.0
gunicorn==20.1.0
h5py==2.10.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==1.1.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
Pillow==8.2.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytesseract==0.3.7
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.4.1
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
